fix ignored "change" event on select elements in IE
In IE, jQuery reuses the click event on select elements to emulate the
"change" event bubbling up.

As shown in http://jsbin.com/ekofa3/14 this breaks when we live-bind to
"click" on any other element in the page. In that case, the event is
processed in liveHandler() and liveFired is set.

Since the event object is reused, liveHandler sees event.liveFired and
ignores the "change" event.

This patch makes sure that when the click event is repurposed liveFired
is cleaned up.